requirements.txt
paho-mqtt==1.6.1
voluptuous>=0.12.2
